Rename Partition -> partition
