--- learning/katas/go/core_transforms/Partition/Partition/cmd/main.go
@@ -1,44 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one or more
-// contributor license agreements.  See the NOTICE file distributed with
-// this work for additional information regarding copyright ownership.
-// The ASF licenses this file to You under the Apache License, Version 2.0
-// (the "License"); you may not use this file except in compliance with
-// the License.  You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package main
-
-import (
-	"context"
-	"github.com/apache/beam/sdks/go/pkg/beam"
-	"github.com/apache/beam/sdks/go/pkg/beam/log"
-	"github.com/apache/beam/sdks/go/pkg/beam/x/beamx"
-	"github.com/apache/beam/sdks/go/pkg/beam/x/debug"
-	"partition/pkg/task"
-)
-
-func main() {
-	ctx := context.Background()
-
-	p, s := beam.NewPipelineWithRoot()
-
-	input := beam.Create(s, 1, 2, 3, 4, 5, 100, 110, 150, 250)
-
-	output := task.ApplyTransform(s, input)
-
-	debug.Printf(s, "Number > 100: %v", output[0])
-	debug.Printf(s, "Number <= 100: %v", output[1])
-
-	err := beamx.Run(ctx, p)
-
-	if err != nil {
-		log.Exitf(context.Background(), "Failed to execute job: %v", err)
-	}
-}
