+++ learning/katas/go/core_transforms/partition/lesson-info.yaml
@@ -17,19 +17,5 @@
 # under the License.
 #
 
-type: edu
-files:
-- name: test/task_test.go
-  visible: false
-- name: cmd/main.go
-  visible: true
-- name: go.mod
-  visible: false
-- name: pkg/task/task.go
-  visible: true
-  placeholders:
-  - offset: 946
-    length: 103
-    placeholder_text: TODO()
-- name: go.sum
-  visible: false
+content:
+- partition
