+++ learning/katas/go/core_transforms/partition/lesson-remote-info.yaml
