+++ learning/katas/go/core_transforms/partition/partition/pkg/task/task.go
