+++ learning/katas/go/core_transforms/partition/partition/task-info.yaml
@@ -17,5 +17,15 @@
 # under the License.
 #
 
-content:
-- Partition
+type: edu
+files:
+- name: test/task_test.go
+  visible: false
+- name: cmd/main.go
+  visible: true
+- name: pkg/task/task.go
+  visible: true
+  placeholders:
+  - offset: 946
+    length: 103
+    placeholder_text: TODO()
