+++ learning/katas/go/core_transforms/partition/partition/task.md
