+++ learning/katas/go/core_transforms/partition/partition/test/task_test.go
@@ -13,18 +13,36 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-module partition
+package test
 
-go 1.14
-
-require (
-	github.com/apache/beam v2.22.0+incompatible
-	github.com/golang/protobuf v1.4.2 // indirect
-	github.com/google/go-cmp v0.4.1 // indirect
-	github.com/googleapis/gax-go v1.0.3 // indirect
-	golang.org/x/net v0.0.0-20200602114024-627f9648deb9 // indirect
-	golang.org/x/sys v0.0.0-20200610111108-226ff32320da // indirect
-	google.golang.org/api v0.26.0 // indirect
-	google.golang.org/genproto v0.0.0-20200611194920-44ba362f84c1 // indirect
-	google.golang.org/grpc v1.29.1 // indirect
+import (
+	"beam.apache.org/learning/katas/core_transforms/partition/partition/pkg/task"
+	"github.com/apache/beam/sdks/go/pkg/beam"
+	"github.com/apache/beam/sdks/go/pkg/beam/testing/passert"
+	"github.com/apache/beam/sdks/go/pkg/beam/testing/ptest"
+	"testing"
 )
+
+func TestApplyTransform(t *testing.T) {
+	p, s := beam.NewPipelineWithRoot()
+	tests := []struct {
+		input beam.PCollection
+		want [][]interface{}
+	}{
+		{
+			input: beam.Create(s, 1, 2, 3, 4, 5, 100, 110, 150, 250),
+			want: [][]interface{}{
+				{110, 150, 250},
+				{1, 2, 3, 4, 5, 100},
+			},
+		},
+	}
+	for _, tt := range tests {
+		got := task.ApplyTransform(s, tt.input)
+		passert.Equals(s, got[0], tt.want[0]...)
+		passert.Equals(s, got[1], tt.want[1]...)
+		if err := ptest.Run(p); err != nil {
+			t.Error(err)
+		}
+	}
+}
